requirements.txt
fastapi
uvicorn
sqlalchemy
openai
requests
python-telegram-bot